requirements.txt
web3>=7.0.0
requests>=2.32.0
python-dotenv>=1.0.0
eth-account>=0.13.0